Notable upstream additions: VSA packed mixed-modality sparse attention
for MiniMax H3 (#1695), rank-reduced AdaLN pruned-model option (#1699),
H3 pipeline cleanup (#1698) and SFT pipeline (#1688), MMAudio V2A
(#1622), profiler fixes (#1691), two repo-wide yapf passes.

One conflict, fastvideo/attention/backends/flash_attn.py: upstream
generalized the masked flash_attn_no_pad call from causal=False to
causal=self.causal; our fork adds the FA4-FP8 elif branch on the
adjacent lines. Took both — upstream's causal fix on the shared line,
our _fa4_fp8_active branch preserved.

+++ fastvideo/tests/ops/quantization/test_nvfp4_purge.py
@@ -46,8 +46,7 @@ class _FakeLinear(nn.Module):
 
 def _fake_quantize(weight, global_sf, sfLayout=None, do_shuffle=False):
     out_dim, in_dim = weight.shape[0], weight.shape[-1]
-    return (torch.zeros(out_dim, in_dim // 2, dtype=torch.int8), torch.zeros(out_dim, in_dim // 16,
-                                                                             dtype=torch.uint8))
+    return (torch.zeros(out_dim, in_dim // 2, dtype=torch.int8), torch.zeros(out_dim, in_dim // 16, dtype=torch.uint8))
 
 
 @pytest.fixture(autouse=True)
